+++ CHANGELOG.md
@@ -59,6 +59,7 @@ The format is a modified version of [Keep a Changelog](https://keepachangelog.co
 - Guard download cache serialization [@mrissaoussama](https://github.com/mrissaoussama) [#346](https://github.com/tsundoku-otaku/tsundoku/pull/346)
 - Paginated library export to prevent potential OOM, added cancel to notif [@mrissaoussama](https://github.com/mrissaoussama) [#270](https://github.com/tsundoku-otaku/tsundoku/pull/270)
 - Correct import stats when app is restarted [@mrissaoussama](https://github.com/mrissaoussama) [#272](https://github.com/tsundoku-otaku/tsundoku/pull/272)
+- Consistent JS/custom tagging, fixed duplicate check, add options in library settings [@mrissaoussama](https://github.com/mrissaoussama) [#359](https://github.com/tsundoku-otaku/tsundoku/pull/359)
 - Fix for an Injekt related exception tied to performance improvements, should keep performance [@mrissaoussama](https://github.com/mrissaoussama) [#352](https://github.com/tsundoku-otaku/tsundoku/pull/352)
 - Properly fetch details and chapters through new getMangaUpdate [@mrissaoussama](https://github.com/mrissaoussama) [#363](https://github.com/tsundoku-otaku/tsundoku/pull/363)
 - Fix download queue restart preference [@mrissaoussama](https://github.com/mrissaoussama) [#297](https://github.com/tsundoku-otaku/tsundoku/pull/297)